+++ jina/docker/helper.py
@@ -1,14 +1,13 @@
 __copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-import time
 import base64
+import time
 from pathlib import Path
+from typing import Dict, Union, List
 
 from .. import __binary_delimiter__
 from ..logging import JinaLogger
-from pkg_resources import resource_stream
-from typing import Dict, Union, List
 
 
 def _encode(clear, key=__binary_delimiter__.decode()):
@@ -51,19 +50,19 @@ class Waiter:
         self.logger = JinaLogger(self.__class__.__name__)
         self._seconds = seconds
         self._message = message
-        
+
     def __enter__(self):
         self.logger.info(f'waiting for {self._seconds} seconds {self._message}')
         self._wait_until = time.time() + self._seconds
         return self
-    
+
     @property
     def is_time_up(self):
         return time.time() > self._wait_until
-    
+
     def sleep(self, seconds=5):
         self.logger.debug(f'sleeping for {seconds} seconds')
         time.sleep(seconds)
-    
+
     def __exit__(self, type, value, traceback):
         self.logger.debug(f'took {self._seconds} seconds!')
